requirements.txt
agate==1.6.1
agate-dbf==0.2.0
agate-excel==0.2.2
agate-sql==0.5.3
Babel==2.6.0
beautifulsoup4==4.6.3
csvkit==1.0.3
cycler==0.10.0
dbfread==2.0.7
et-xmlfile==1.0.1
future==0.17.1
isodate==0.6.0
japanize-matplotlib==1.0.5
jdcal==1.4
joblib==0.14.1
kiwisolver==1.1.0
leather==0.3.3
matplotlib==3.2.1
numpy==1.18.2
openpyxl==2.5.11
pandas==1.0.3
parsedatetime==2.4
Pillow==7.1.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-slugify==1.2.6
pytimeparse==1.1.8
pytz==2018.7
scikit-learn==0.22.2.post1
scipy==1.4.1
selenium==3.14.1
six==1.11.0
sklearn==0.0
SQLAlchemy==1.2.14
Unidecode==1.0.23
urllib3==1.24
xlrd==1.1.0
xlsx2csv==0.7.4
